Fix for compatability problems between osx and linux
More specifically, bash and coreutils has major compatability problems on osx as it hasn't been updated in at least a decade due to lisencing problems. Instead of hacking around this, I chose to upgrade bash and coreutils on osx on my default setup via brew, this is harmless as it does not overwrite the default bash and osx.